обработка ответа со null значениями и ловим превышения лимита запросов

+++ app/src/main/java/com/example/binlisttesttask/core/data/database/entity/CountryEntity.kt
@@ -17,7 +17,7 @@ data class CountryEntity(
     @ColumnInfo(name = "currency")
     val currency: String?,
     @ColumnInfo(name = "latitude")
-    val latitude: Int?,
+    val latitude: Float?,
     @ColumnInfo(name = "longitude")
-    val longitude: Int?
+    val longitude: Float?
 )
