+++ app/src/main/java/com/example/binlisttesttask/feature/main/data/models/CountryDto.kt
@@ -6,6 +6,6 @@ class CountryDto (
     val name: String? = null,
     val emoji: String? = null,
     val currency: String? = null,
-    val latitude: Int?    = null,
-    val longitude: Int?    = null
+    val latitude: Float?    = null,
+    val longitude: Float?    = null
 )
