+++ app/src/main/java/com/example/binlisttesttask/feature/main/domain/models/Country.kt
@@ -6,6 +6,6 @@ class Country (
     val name: String? = null,
     val emoji: String? = null,
     val currency: String? = null,
-    val latitude: Int?    = null,
-    val longitude: Int?    = null
+    val latitude: Float?    = null,
+    val longitude: Float?    = null
 )
